[updated] :root 정의, 중고 상세페이지 UI 부분 살짝 리팩토링, 폰트 적용
